feat(compaction): add AI-driven compression and semantic pruning
Configurable compression levels with importance-based pruning for
conversation history. AI-driven semantic analysis to identify and
compress low-value exchanges while preserving critical context.

+++ crates/forge_app/src/compression/mod.rs
@@ -32,7 +32,7 @@ pub fn compress(context: impl Into<Context>, config: &Compact) -> (Context, Comp
     let mut report = CompressionReport::default();
 
     let total_tokens = ctx.token_count_approx();
-    let budget = config.token_threshold.unwrap_or(80_000) as usize;
+    let budget = config.token_threshold.unwrap_or(80_000);
     let level = config.context_compression_level;
 
     if total_tokens <= budget {
